332: Add context map
This is based on the domain driven design concept that in order to represent
things well an application should be split into various contexts. This helps
avoid creating objects that try to do too much an become unmaintable.

With this concept of bounded contexts in place i.e. the staging context and the
points context an explicit definition of how to translate from one context to
another needs to be introduced - enter context map.

In our use case we will only ever map from the staging schema to the app schema
therefore the context will only be applied to staging objects and produce app
objects.