332: Test that any missing argument throws an error (#9)
As javascript gives so little assurance about object types we need to ensure
that our objects are as we expect when they are constructed. This can lead to a
lot of unit tests with duplicate code.

This is an approach which reduces the amount of code required to test all of the
parameters.