Feature/wmt 414/display arms assessments (#25)
* 414: Add ARMS assessments to case type weightings

* Update case-type-weightings to accept required ARMS weightings

* 414: Add arms fields to the probation rules mapper

Add additional unit tests which highlighted a bug with the license and community
last 16 weeks fields

* 414: Update workload points calculations for ARMS assessments

* 414: Move SDR, SDR Conversion and PAROMs into workload points calculation

The probation rules should be responsible for as much of the workload points
calculations as possible. This will contain the business logic and reduce the
complexity fo the calculate points worker.

As this is a breaking change the probation rules version will be increased to
2.0.

* 414: Move calculateWeighting method from worker into probation rules

* 414: Move calculate weighting function into tier points calculations

The calculate weighting function is only used in the `calculate-points-for-tier`
file. Therefore there was no need to have it outside of the class. It was also
at risk of being used incorrectly by another file.